.github/tests/acceptance/test_blink_led_heartbeat/requirements.txt
pytest
rpi-lgpio
pytest-timeout
numpy
scipy